+++ LevelUp/Domain/GameData.cs
@@ -1,8 +1,10 @@
+using LevelUp.Domain.Books;
 using LevelUp.Domain.Bosses;
 using LevelUp.Domain.Habits;
 using LevelUp.Domain.Milestones;
 using LevelUp.Domain.Projects;
 using LevelUp.Domain.Quests;
+using LevelUp.Domain.Wallet;
 using CharacterModel = LevelUp.Domain.Character.Character;
 
 namespace LevelUp.Domain;
@@ -20,4 +22,8 @@ public class GameData
     public List<Milestone> Milestones { get; set; } = [];
 
     public List<BossEncounter> Bosses { get; set; } = [];
+
+    public List<Book> Books { get; set; } = [];
+
+    public List<WalletTransaction> WalletTransactions { get; set; } = [];
 }